Bugfix: Added changes to update prompt for webapps (#306)
The PR makes the following changes:

1. Updates the request object to include a client variable to capture
information on which client is sending the request. This information
will be used to update the prompt & also for capturing metrics on
requests per client type. Currently there are two client types - slack &
webapp
2. For each of the client types - slack and webapp, we add conditions to
update the relevant request variable
3. When the request is received from a webapp, then the prompt to LLM is
updated to specify that we must return the result in markdown format